==> scripts/EL/00-retrieve-metadata-export-annotations.ipynb <==
# 00 - Retrieve elements and metadata on Arkindex

This notebook aims to retrieve data on Arkindex using Arkindex API. More precisly : 
1. retrieve the considered registers pages
2. retrieve the content of the page covers
3. retrieve the annotations of the tables and if needed create inference-like jsons

import pandas as pd
import glob
import json
from arkindex import ArkindexClient

import sys
import os
# Access to the utils directory
current_dir = os.getcwd()
utils_dir = os.path.join(current_dir, '..', 'utils')
sys.path.append(utils_dir)

### Parameters

NAME = "LHAY"
FOLDER = f"/home/STual/DAN-cadastre/inference/{NAME}"
SAVE_FOLDER = f'/home/STual/DAN-cadastre/data/{NAME}'

PAGE_ELEMENTS_SLUGS = ["page_add","page_add_to_cor"] #Marolles : ["page", "page_marolles"]
ARKINDEX_FOLDER = "2ec47da6-116b-4f53-bd93-da7aac699a4a" #marolles : "2a045b63-2866-4292-a168-87d342be800b"
PAGE_CLASSES = ["ets_couv","ets_tab_p1","ets_tab_p2","ets_recap_inter","ets_resume"]

cli = ArkindexClient('cdf9b285c9da6e82cddbf18aa070dc471a13facb')

## 1. Retrieve pages infos

elements = cli.request('ListElementChildren',
                   id=ARKINDEX_FOLDER,type=PAGE_ELEMENTS_SLUGS[0],
                   recursive=True,folder=False,page_size=500,page=1,with_classes=True,
                )
for slug in PAGE_ELEMENTS_SLUGS[1:]:
    elements2 = cli.request('ListElementChildren',
                           id=ARKINDEX_FOLDER,type=slug,
                           recursive=True,folder=False,page_size=500,page=1,with_classes=True,
                        )
    elements["results"].extend(elements2["results"])
print(f"{len(elements['results'])} elements of type {', '.join(PAGE_ELEMENTS_SLUGS)} have been retrieved")

Now, we create a table with the informations of each element (pages) and save them to a CSV.

from arkindex_utils import listelements_withclass_json_to_table

df_commune = listelements_withclass_json_to_table(elements,PAGE_CLASSES,["ets_blank"],["is_blank"])
df_commune = df_commune[df_commune["classe"].isin(["ets_couv", "ets_tab_p1"])]
df_commune['image_index'] = df_commune['image_index'].astype('int')
df_commune.sort_values(by=['image_index'],inplace=True)
df_commune.to_csv(f'{SAVE_FOLDER}/{NAME}_pages.csv',index=False)

df_commune[0:4]

## 2. Create page covers info files

EXTRACT_COVERS_CONTENT = True

COVERS_CSV_COLUMNS = ["page_uuid","element_uuid","element_type","polygon","transcription_uuid","transcription"]

if EXTRACT_COVERS_CONTENT:
    df_commune_couv = df_commune[df_commune["classe"] == "ets_couv"]
    areas = []
    for _, row in df_commune_couv.iterrows():
        #List the areas
        results = cli.request('ListElementChildren',
                           id=row["id"],
                           with_transcriptions=True,page=1,page_size=500)
        for l in results["results"]:
            area_id = l["id"]
            type_ =  l["type"]
            polygon = l["zone"]["polygon"]
            transcription_id = l["transcriptions"][0]["id"]
            transcription = l["transcriptions"][0]["text"]
            areas.append([row["id"], area_id, type_, polygon, transcription_id, transcription])
            
    #Save data
    areas_df = pd.DataFrame(areas,columns=COVERS_CSV_COLUMNS)
    areas_df.to_csv(f'{SAVE_FOLDER}/{NAME}_couv.csv', index=False)
    display(areas_df[0:6])

## 3. Create inference-like data from Arkindex
### 3.1 Retrieve table lines transcriptions and entities

Create a CSV with the transcriptions of each table line of the pages listes before. The CSV has the following header : ["page_uuid","line_uuid","polygon","transcription_uuid","transcription","entities"]

DOWNLOAD_TRANSCRIPTIONS = False

if DOWNLOAD_TRANSCRIPTIONS:
    df_commune_tab = df_commune[df_commune["classe"] == "ets_tab_p1"]
    lines = []
    for _, row in df_commune_tab.iterrows():
        #List the lines
        results = cli.request('ListElementChildren',
                           id=row["id"],
                           with_transcriptions=True,page=1,page_size=500)
        
        for l in results["results"]:
            try:
                if "tab_header" not in l["type"]:
                    line_id = l["id"]
                    polygon = l["zone"]["polygon"]
                    transcription_id = l["transcriptions"][0]["id"] #Il ne retrouve pas les transcriptions, pourquoi ? (contraindre les types de lignes ?)
                    transcription = l["transcriptions"][0]["text"]
                    #List the entities
                    entities = cli.request("ListTranscriptionEntities",id=transcription_id)
                    ls_ent = []
                    for e in entities["results"]:
                        js = {"entity_uuid" : e["entity"]["id"], "entity_text" : e["entity"]["name"], "entity_type" : e["entity"]["type"]["name"], "offset" : e["offset"], "length": e["length"]}
                        ls_ent.append(js)
                    lines.append([row["id"],line_id,polygon,transcription_id,transcription,ls_ent])
            except:
                print(l) 

if DOWNLOAD_TRANSCRIPTIONS:
    lines_df = pd.DataFrame(lines,columns=["page_uuid","line_uuid","polygon","transcription_uuid","transcription","entities"])
    lines_df[0:4]

if DOWNLOAD_TRANSCRIPTIONS:
    lines_df.to_csv(f'{SAVE_FOLDER}/{NAME}_tab_lines.csv', index=False) 

### 3.2 Create json similar to DAN inference

from string_utils import TableValuesPostTreatment 

CREATE_JSONS = False

if CREATE_JSONS:
    df_pages = pd.read_csv(f'{SAVE_FOLDER}/{NAME}_pages.csv')
    df_pages = df[df["classe"] == "ets_tab_p1"]
    df_entities = pd.read_csv(f'{SAVE_FOLDER}/{NAME}_tab_lines.csv')

if CREATE_JSONS:

    for _, row in df_pages.iterrows():
        page_json = {}
        name = row["id"]
        page_json["text"] = ""
        page_json["confidences"] = {}
        page_json["confidences"]["total"] = 1.0
        page_json["language_model"] = {}
        page_json["objects"] = []
        page_json["attention_gif"] = ""
        page_json["entities"] = ""
        subdf = df_entities[df_entities["page_uuid"] == name]
        new_ents = []
        for __, row2 in subdf.iterrows():
            obj = {"confidence": 1.0, "polygon":row2["polygon"],"text": row2["transcription"],"text_confidence": 1.0}
            page_json["objects"].append(obj)
            ents = row2["entities"].replace("\'entity_text\'",'"entity_text"').replace("\'entity_uuid\'",'"entity_uuid"').replace("\'entity_type\'",'"entity_type"').replace("\'offset\'",'"offset"').replace("\'length\'",'"length"').replace(": '",': "').replace("',",'",')
            ents = json.loads(ents)
            line_en = {}
            for e in ents:
                t = ''
                if e["entity_type"] == "lieu-dit":
                    t = "Ⓓ"
                elif e["entity_type"] == "ancien_numero_parcelle":
                    t = "Ⓐ"
                elif e["entity_type"] == "ancienne_nature":
                    t = "Ⓑ"
                elif e["entity_type"] == "identite":
                    t = "Ⓒ"
                elif e["entity_type"] == "nature":
                    t = "Ⓔ"
                elif e["entity_type"] == "numero_parcelle":
                    t = "Ⓕ"
                elif e["entity_type"] == "numero_proprietaire":
                    t = "Ⓖ"
                line_en.update({t:{"text": e["entity_text"],
                    "offset": e["offset"],
                    "length": e["length"]}})
            new_ents.append(line_en)
        page_json["entities"] = new_ents
        #Deal with items and dittos
        page_json["entities"] = TableValuesPostTreatment.process_idem_ditto_replacements(page_json["entities"], SPECIAL_VALUE="MISSING", 
                                        idem_list=['§', 'Ø', 'id', 'idem', 'le meme', 'la meme', 'les memes'],
                                        ditto_list=['☼'])
        json_path = os.path.join(SAVE_FOLDER,"data","LHAY","ets_tab_p1",row['id'] + '.json')
        #Save the jsons
        with open(json_path,'w', encoding='utf-8') as f:
            json.dump(page_json, f, ensure_ascii=False, indent=4)
