==> scripts/EL/11-validation-DAN-NER.ipynb <==
# 11 - Export validation CSV to create ground-truth

import pandas as pd
import glob
import json

import sys
import os
# Access to the utils directory
current_dir = os.getcwd()
utils_dir = os.path.join(current_dir, '..', 'utils')
sys.path.append(utils_dir)

from string_utils import TableValuesPostTreatment, NormalizeText

ROOT = "/home/STual/DAN-cadastre/inference/LHAY/"
RES = "/home/STual/DAN-cadastre/data/LHAY/ets_tab_p1"
JSONS = glob.glob(ROOT + '*.json')

## 1. Create CSVs

tabs = []
entities = ["Ⓒ","Ⓓ","Ⓔ","Ⓕ","Ⓖ"]
csv_columns = ["page_uuid","line_uuid","line_detection_order","line_true_order","line_status","taxpayer_num","taxpayer_cell","taxpayer_interpreted","name","firstnames","birthname","activities","addresses","titles","familystatus","plot_number","plot_address","plotnature","comment"]

for JSON in JSONS:
    with open(JSON,'r',encoding='utf8') as f:
        page = json.load(f)
    page_uuid = JSON.replace(ROOT,"").replace('.json','')

    tab = []
    counter = 0
    for line in page["entities"]:
        keys = list(line.keys())
        fkeys = keys[:-1]
        line_uuid = line["uri"]
        
        key = 'Ⓖ'
        if 'Ⓖ' in fkeys:#Numéro dans la table alphabétique
            taxpayer_num_cell = line[key]["text"]
        else:
            taxpayer_num_cell = ""

        key = "Ⓒ"
        if "Ⓒ" in fkeys: #Taxpayer
            taxpayer_cell = line[key]["text"]
            taxpayer_cell_i = line[key]["interpreted_text"] #A propager dans le tableau
            ner = line[key]["ner_interpreted_text_grouped"] #Si liste, séparer par des |. A propager dans le tableau si besoin
            name = ner["name"]
            firstnames = ner["firstnames"]
            birthname = ner["birthname"]
            if len(ner["activity"]) > 0:
                activities = "|".join(ner["activity"])
            else:
                activities = ""
            if len(ner["address"]) > 0:
                addresses = "|".join(ner["address"])
            else:
                addresses = ""
            if len(ner["title"]) > 0:
                titles = "|".join(ner["title"])
            else:
                titles = ""
            if len(ner["familystatus"]) > 0:
                familystatus = "|".join(ner["familystatus"])
            else:
                familystatus = ""
        else:
            taxpayer_cell, taxpayer_cell_i = "", ""
            
        key = "Ⓕ"
        if key in fkeys:
            plot_num = line[key]["text"]
        else:
            plot_num = ""
            
        key = "Ⓓ" 
        if key in fkeys:
            plot_address = line[key]["text"]
        else:
            plot_address = ""
            
        key = "Ⓔ"
        if key in fkeys:
            plot_nature = line[key]["text"]
        else:
            plot_nature = ""
            
        csv_line = [page_uuid,line_uuid,str(counter),"","",taxpayer_num_cell,taxpayer_cell,taxpayer_cell_i,name,firstnames,birthname,activities,addresses,titles,familystatus,plot_num,plot_address,plot_nature,""]
        assert len(csv_line) == len(csv_columns)
        tab.append(csv_line)
        counter+=1
    tabs.append(tab)

for tab in tabs:
    df = pd.DataFrame(tab,columns=csv_columns)
    page_uuid = df.iloc[0]["page_uuid"]
    df.to_csv(RES + "/" + page_uuid + '.csv', index=False) 

df
