==> scripts/EL/20-entity-linking-with-NIL-option.ipynb <==
# 20 - Entity linking with NIL option : example of the plot natures
* This nobook propose methods to perform entity linking between historical tables cells and existing entities in a domain KG.
* Developped with the example of land plots natures (already 160 natures in the ontology).

import pandas as pd
import rdflib
from rdflib.namespace import SKOS, RDF, RDFS
import glob

import sys
import os
# Access to the utils directory
current_dir = os.getcwd()
utils_dir = os.path.join(current_dir, '..', 'utils')
sys.path.append(utils_dir)

## 1. Retrieve candidates from the domain ontology

from rdf import skosConceptUtil

CSV_PATH = '/home/STual/DAN-cadastre/data/ONTO/SKOS_NATURE.csv'
skos_graph = skosConceptUtil()
skos_graph.init_nature_concepts_from_csv(CSV_PATH,'/home/STual/DAN-cadastre/data/ONTO/')

# Create a graph
g = rdflib.Graph()

# Parse the Turtle file into the graph
g.parse("/home/STual/DAN-cadastre/data/ONTO/natures.ttl", format="ttl")

# Define a SPARQL query that uses the vocabularies as variables
query = """
PREFIX skos: <http://www.w3.org/2004/02/skos/core#>
PREFIX rdfs: <http://www.w3.org/2000/01/rdf-schema#>

SELECT ?subject ?label
WHERE {
  ?subject a skos:Concept .
  {?subject skos:prefLabel ?label .} UNION {?subject skos:altLabel ?label .} UNION {?subject skos:hiddenLabel ?label .}
  FILTER(LANG(?label) = "fr")
}
"""
# Execute the query on the graph
results = g.query(query)

results_ls = []
# Print the results
for row in results:
    results_ls.append([str(row.subject),str(row.label)])

results_ls[0:5]

labels_ = [x[1] for x in results_ls]
print(labels_)

## 2. Retrieve mentions

#FOLDER = "/home/STual/DAN-cadastre/inference/training120325_config2025_prod_2000epochs-DEV"
FOLDER = "/home/STual/DAN-cadastre/inference/LHAY"
JSONS = glob.glob(FOLDER + '/*.json')
print(len(JSONS))

* *Idem* and *ditto* have to be replaced by their concrete value (at the page level)

import json
import glob
from string_utils import TableValuesPostTreatment

chars_to_split = ["Ⓐ", "Ⓑ", "Ⓒ", "Ⓓ", "Ⓔ", "Ⓕ", "Ⓖ"]

# Extract cell contents from line transcription
#TableValuesPostTreatment.update_jsons_with_objects_details(FOLDER, ["Ⓐ", "Ⓑ", "Ⓒ", "Ⓓ", "Ⓔ", "Ⓕ", "Ⓖ"])

KEY = "Ⓔ"
IX = 4
distinct_natures = []
#for j in JSONS:
with open(JSONS[IX]) as f:
    page_name = JSONS[IX].replace(FOLDER+'/','').replace('.json','')
    print(page_name)
    page = json.load(f)
    cells = page["entities"]
    
    for c in cells:
        keys = c.keys()
        if KEY in keys:
            print(c[KEY])
            distinct_natures.append(c[KEY]['interpreted_text'])
        else:
            print('MISSING')
distinct_natures = list(set(distinct_natures))
distinct_natures = [d for d in distinct_natures if d is not None]
print(distinct_natures)

## 3. Entity linking
* Entity-linking has to consider NIL entity linking as all the mentions in historical tables have no corresponding concept in the ontology.

### 3.1 Edit distance based similarities

#### Gestalt pattern matching (Difflib)

from linking_utils import EditDistanceBasedSimilarity

remove_chars_regex = '→()↑↓×.,!?;:-@#$%^&*'
replacement_char = ' '
split_chars = '→, .;'
matched_natures =  EditDistanceBasedSimilarity.EL_gestaltpatternmatching(distinct_natures, labels_, remove_chars_regex, replacement_char, split_chars, top_k=1, threshold=0.7)

### 3.2 Embedding similarity

from linking_utils import LinkingUtils, EmbeddingSimilarity
from sentence_transformers import SentenceTransformer

from linking_utils import LinkingUtils, EmbeddingSimilarity
from sentence_transformers import SentenceTransformer

embedder_name = "all-MiniLM-L6-v2"
embedder = SentenceTransformer(embedder_name)

remove_chars_regex = '→()↑↓×.,!?;:-@#$%^&*'
replace_regex = "→()↑↓×."
split_regex = "., "

embedding_results = EmbeddingSimilarity.EL_embedding_cosinedistance_matching(distinct_natures, labels_, embedder, remove_chars_regex, replacement_char, split_chars, top_k=1, threshold=0.8)
print(embedding_results)

## 4. Add results to jsons

from linking_utils import LinkingUtils

uris_values = LinkingUtils.associate_matches_with_values(distinct_natures, matched_natures, results_ls, 0)
uris_values

## 5. Main

JSONS = sorted(glob.glob(FOLDER + '/*.json'))
KEY = "Ⓔ"
METHOD = "editdistancebased"

for JSON in JSONS:
    print()
    distinct_natures = []
    with open(JSON) as f:
        page_name = JSON.replace(FOLDER,'').replace('.json','').replace('/','')
        print(page_name)
        page = json.load(f)
        cells = page["entities"]
        
        for c in cells:
            keys = c.keys()
            if KEY in keys:
                distinct_natures.append(c[KEY]['interpreted_text'])
                
    distinct_natures = list(set(distinct_natures))
    distinct_natures = [d for d in distinct_natures if d is not None]
    
    remove_chars_regex = '→()↑↓×.,!?;:-@#$%^&*±'
    replacement_char = ' '
    split_chars = '→, .;'
    if METHOD == "embeddingbased":
        embedding_results = EmbeddingSimilarity.EL_embedding_cosinedistance_matching(distinct_natures, labels_, embedder, remove_chars_regex, replacement_char, split_chars, top_k=1, threshold=0.8)
    elif METHOD == "editdistancebased":
        matched_natures =  EditDistanceBasedSimilarity.EL_gestaltpatternmatching(distinct_natures, labels_, remove_chars_regex, replacement_char, split_chars, top_k=1, threshold=0.7)
    
    uris_values = LinkingUtils.associate_matches_with_values(distinct_natures, matched_natures, results_ls, 0)
    for c in page["entities"]:
        keys = c.keys()
        if KEY in keys :
            if c[KEY]["interpreted_text"] is not None:
                c[KEY]['uris'] = uris_values[c[KEY]["interpreted_text"]]
        
    with open(JSON, 'w') as f:
        json.dump(page, f, indent=4)
