==> scripts/EL/61-create-gold-standard-jsons.ipynb <==
import glob
import json
import pandas as pd

import sys
import os
# Access to the utils directory
current_dir = os.getcwd()
utils_dir = os.path.join(current_dir, '..', 'utils')
sys.path.append(utils_dir)
from string_utils import TableValuesPostTreatment, NormalizeText

## 1. Init Gold Standard jsons

PAGE_CSVS = sorted(glob.glob("/home/STual/DAN-cadastre/data/LHAY/corrected_updated/1*.csv"))
ENTITIES_CSV = sorted(glob.glob("/home/STual/DAN-cadastre/data/LHAY/corrected_updated/*_entities.csv"))
SAVE_FOLDER = "/home/STual/DAN-cadastre/data/LHAY/corrected_jsons"

len(PAGE_CSVS)

def list_in_cell(key,sep="|",join_sep = ", "):

    if pd.notna(row[key]) and sep in row[key]:
        return join_sep.join(row[key].split(sep)),
    elif pd.notna(row[key]) and "|" not in row[key]:
        return [row[key]]
    else:
        return []

for PAGE in PAGE_CSVS:
    df_ = pd.read_csv(PAGE,encoding='utf8')
    df = df_[df_["line_status"].isin(["VALIDE", "MANQUANTE"])]
    page_json = {"text":"","confidences":{"total":1.0},"language_model":"","objects":[],"entities":[]}
    page_uuid = df.iloc[0]["page_uuid"]
    page_text = ""
    for _, row in df.iterrows():
        if row["line_uuid"] != "28215bed-0304-4dd6-8360-2fd0f81f2b43":
            line_ent = {}
            line = ""
            if pd.notna(row["plot_address"]) and row["plot_address"] != "":
                line_ent["Ⓓ"] = {}
                line_ent["Ⓓ"]["text"] = row["plot_address"]
                line_ent["Ⓓ"]["offset"] = "NC"
                line_ent["Ⓓ"]["length"] = "NC"
                line_ent["Ⓓ"]["interpreted_text"] = ""
                line_ent["Ⓓ"]["uris"] = "http://rdf.geohistoricaldata.org/id/landmark/" + row["landmark_uuid"]
                line += "Ⓓ" + row["plot_address"] + " "
                
            if pd.notna(row["taxpayer_num"]) and row["taxpayer_num"] != "":
                line_ent["Ⓖ"] = {}
                if isinstance(row["taxpayer_num"],float):
                    line_ent["Ⓖ"]["text"] = str(int(row["taxpayer_num"]))
                elif isinstance(row["taxpayer_num"],int):
                    line_ent["Ⓖ"]["text"] = str(row["taxpayer_num"])
                line_ent["Ⓖ"]["offset"] = "NC"
                line_ent["Ⓖ"]["length"] = "NC"
                line_ent["Ⓖ"]["interpreted_text"] = ""
                line += "Ⓖ" + str(row["taxpayer_num"]) + " "
    
            if pd.notna(row["taxpayer_cell"]) and row["taxpayer_cell"] != "":
                line_ent["Ⓒ"] = {}
                line_ent["Ⓒ"]["text"] = row["taxpayer_cell"]
                line_ent["Ⓒ"]["offset"] = "NC"
                line_ent["Ⓒ"]["length"] = "NC"
                line_ent["Ⓒ"]["interpreted_text"] = row["taxpayer_interpreted"]
                line_ent["Ⓒ"]["uris"] = ["http://rdf.geohistoricaldata.org/id/taxpayer/" + row["taxpayer_uuid"]]
                line += "Ⓒ" + row["taxpayer_cell"] + " "
                line_ent["Ⓒ"]["ner_interpreted_text_grouped"] = {}
                
                if pd.notna(row["name"]):
                    line_ent["Ⓒ"]["ner_interpreted_text_grouped"]["name"] = row["name"]
                else:
                    line_ent["Ⓒ"]["ner_interpreted_text_grouped"]["name"] = ""
    
                if pd.notna(row["firstnames"]):
                    if '|' in row["firstnames"]:
                        ls = row["firstnames"].split('|')
                        name = ' et '.join(ls)
                    else:
                        name = row["firstnames"]
                    line_ent["Ⓒ"]["ner_interpreted_text_grouped"]["firstnames"] = name
                else:
                    line_ent["Ⓒ"]["ner_interpreted_text_grouped"]["firstnames"] = ""
    
                if pd.notna(row["birthname"]):
                    if '|' in row["birthname"]:
                        ls = row["birthname"].split('|')
                        bname = ' et '.join(ls)
                    else:
                        bname = row["birthname"]
                    line_ent["Ⓒ"]["ner_interpreted_text_grouped"]["birthname"] = bname
                else:
                    line_ent["Ⓒ"]["ner_interpreted_text_grouped"]["birthname"] = ""
                    
                line_ent["Ⓒ"]["ner_interpreted_text_grouped"]["address"] = list_in_cell("addresses",sep="|",join_sep = ", ")
                    
                line_ent["Ⓒ"]["ner_interpreted_text_grouped"]["familystatus"] = list_in_cell("familystatus",sep="|",join_sep = ", ")
    
                line_ent["Ⓒ"]["ner_interpreted_text_grouped"]["activity"] = list_in_cell("activities",sep="|",join_sep = ", ")
                    
                line_ent["Ⓒ"]["ner_interpreted_text_grouped"]["title"] = list_in_cell("titles",sep="|",join_sep = ", ")
                     
            if pd.notna(row["plot_number"]) and row["plot_number"] != "":
                line_ent["Ⓕ"] = {}
                line_ent["Ⓕ"]["text"] = row["plot_number"]
                line_ent["Ⓕ"]["offset"] = "NC"
                line_ent["Ⓕ"]["length"] = "NC"
                if isinstance(row["plot_number"],str):
                    line_ent["Ⓕ"]["interpreted_text"] = row["plot_number"]
                    line_ent["Ⓕ"]["postcorrected_text"] = row["plot_number"].replace('er','').replace('↓','').replace('↑','').replace('→','')
                elif isinstance(row["plot_number"],float):
                    line_ent["Ⓕ"]["interpreted_text"] = str(int(row["plot_number"]))
                    line_ent["Ⓕ"]["postcorrected_text"] = str(int(row["plot_number"]))
                elif isinstance(row["plot_number"],int):
                    line_ent["Ⓕ"]["interpreted_text"] = str(row["plot_number"])
                    line_ent["Ⓕ"]["postcorrected_text"] = str(row["plot_number"])
                line += "Ⓕ" + str(row["plot_number"]) + " "
    
            if pd.notna(row["plotnature"]) and row["plotnature"] != "":
                line_ent["Ⓔ"] = {}
                line_ent["Ⓔ"]["text"] = row["plotnature"]
                line_ent["Ⓔ"]["offset"] = "NC"
                line_ent["Ⓔ"]["length"] = "NC"
                line_ent["Ⓔ"]["interpreted_text"] = row["plotnature"]
                if pd.notna(row["skosNature"]):
                    line_ent["Ⓔ"]["uris"] = ["http://rdf.geohistoricaldata.org/id/codes/cadastre/plotNature/" + row["skosNature"]]
                line += "Ⓔ" + row["plotnature"] + " "
                
            line_ent["uri"] = row["line_uuid"]
            line_ent["status"] = row["line_status"]
            line_ent["detection_order"] = int(row["line_detection_order"])
            line_ent["true_order"] = int(row["line_true_order"])
            page_text = page_text + '\n' + line 
            page_json["entities"].append(line_ent)
            page_json["objects"].append({"confidence":"NC","polygon":"NC","text":line,"text_confidence":1.0})

    page_json["text"] = page_text
    with open(SAVE_FOLDER + '/' + page_uuid + '.json', "w", encoding="utf-8") as f:
        json.dump(page_json, f, ensure_ascii=False, indent=4)

df

## 2. Post-treat data
* idem
* cells to merge

from string_utils import create_split_dictionary
import uuid

chars_to_split = ["Ⓐ","Ⓑ","Ⓒ","Ⓓ","Ⓔ","Ⓕ","Ⓖ"]
ROOT = "/home/STual/DAN-cadastre/"
JSONS = glob.glob("/home/STual/DAN-cadastre/data/LHAY/corrected_jsons/*.json")

for JSON in JSONS:
    with open(JSON,'r',encoding='utf8') as f:
        page = json.load(f)
    page_uuid = JSON.replace(ROOT,"").replace('.json','')

    #Deal with items and ditto
    page["entities"] = TableValuesPostTreatment.process_idem_ditto_replacements(page["entities"], SPECIAL_VALUE="MISSING", 
                                    idem_list=['§', 'Ø', 'id', 'idem', 'le meme', 'la meme', 'les memes'],
                                    ditto_list=['☼'],to_remove=["uri","detection_order","true_order","status"])
    
    json_path = os.path.join(ROOT, page_uuid + '.json')
    
    with open(json_path,'w', encoding='utf-8') as f:
        json.dump(page, f, ensure_ascii=True, indent=4)
