==> scripts/EL/62-create-gold-standard-resources.ipynb <==
# 02. Create Plot and Documents RDF Resources
* For Gold standard annotations

import pandas as pd
import glob
import json
import uuid
import re
import numpy as np
from rdflib import Graph, URIRef, Literal, RDFS, Namespace, BNode
from rdflib.namespace import SKOS, RDF, RDFS, DCTERMS, XSD

import sys
import os
# Access to the utils directory
current_dir = os.getcwd()
utils_dir = os.path.join(current_dir, '..', 'utils')
sys.path.append(utils_dir)

from string_utils import NormalizeText
from rdf_resources_utils import generate_rdf_resource_section_landmark, generate_rdf_resource_commune_landmark

COMMUNE = "LHAY"
DEP = "94"

ROOT = f"/home/STual/DAN-cadastre/data/{COMMUNE}/corrected_jsons"
DATA = f"/home/STual/DAN-cadastre/data/{COMMUNE}"
JSONS = glob.glob(ROOT + '/*.json')

if not os.path.exists(f'{DATA}/rdf'):
    os.makedirs(f'{DATA}/rdf')

### Retrieve metadata

metadata = pd.read_csv("/home/STual/DAN-cadastre/data/METADATA/details_archives_94_ETS.csv")
metadata = metadata[metadata["Nom_dossier"] == COMMUNE]
metadata

## 1. Create landmarks of type Section and Commune 
* Section : Use covers information.
* Commune : Use archives metadata
### 1.1 Create landmarks of type Section
*TODO : currently, retrieve covers info is done manually. A data template and DAR model to perform IE from covers would be necessary to create a full registers treatment pipeline.*

df_couv = pd.read_csv(f'{DATA}/{COMMUNE}_couv.csv')

df_couv['id'] = df_couv.groupby('page_uuid').cumcount()
# Pivot the table to transform covers elements into attributs of the page covers
pivoted = df_couv.pivot(index='page_uuid', columns='element_type', values='transcription')
pivoted = pivoted.sort_values(by=["ets_couv_id"])
pivoted = pivoted.reset_index()

pivoted["cad_ltype"] = "Section"
pivoted["identifier"] = pivoted["ets_couv_id"].str[-1]
pivoted["label"] = pivoted["ets_couv_id"].str.title() + ' ' + pivoted["ets_couv_name"]
pivoted

The UUID of a Landmark of type section is the cover uuid + the section letter identifier (in case several sections are represented in the same page cover).

#Create a dictionnary for each section, used to create the RDF resources
landmarks_df = pivoted[["cad_ltype","identifier","label","page_uuid"]]
landmarks_dict = []
for _, row in landmarks_df.iterrows():
    rdf_entities = {"uuid":row["page_uuid"],"cad_ltype":row["cad_ltype"],"identifier":row["identifier"],"label":row["label"],"source_uuid":row["page_uuid"]}
    landmarks_dict.append(rdf_entities)
landmarks_dict[0]

g, uri_dict = generate_rdf_resource_section_landmark(landmarks_dict, DEP, COMMUNE)

### 1.2 Add landmark of type commune into the graph
This inclue the creation of the **Landmark** of type **Commune** and of the **LandmarkRelations** between commune and its sections (of LRTYPE Within).

commune_metadata = metadata[["Departement","Commune","Nom_dossier"]]
commune_metadata = commune_metadata.drop_duplicates()
commune_metadata

rdf_entity = {"label":commune_metadata.iloc[0]["Commune"],"departement_num":DEP,"commune_code":COMMUNE}

#Add the landmark commune to the graph
generate_rdf_resource_commune_landmark(g, rdf_entity)

# Print the RDF graph in Turtle format
g.serialize(destination=f"{DATA}/rdf_gold/landmarks.ttl", format="turtle")

## 2. Create RDF resources of sources

### 2.1 Registers

metadata_simple = metadata[["Commune","Cote","Nom_dossier","Date_doc","Date_infos","Tag_groupe_doc"]]
metadata_simple

# Ouvrir le df des pages => prévoir mapping entre les uuid et les noms des cotes d'archives
df_pages = pd.read_csv(f'{DATA}/{COMMUNE}_pages.csv')
#df_pages = df_all_pages[df_all_pages["classe"] == "ets_tab_p1"]
df_pages[0:5]

from string_utils import NormalizeText
from rdf_resources_utils import arkindex_class_to_skosconcept, ets_type

def cote_page_parser(cote_page):
    """
    Parser of numeric references of archival documents of Val-de-Marne (94)
    As FRAD094_3P_000387_01 or eventually FRAD094_3P_000387_01_0042_P1 (for par of double page)
    """
    cote_parts = cote_page.split("_")
    if len(cote_parts) == 6 and cote_page[-2] == 'P':
        archives, serie, item, num_folder, page, part_page = cote_parts
        archives_ = archives.replace('FRAD','')
        archives = NormalizeText.remove_leading_zeros(archives_)
        item = NormalizeText.remove_leading_zeros(item)
        num_folder = NormalizeText.remove_leading_zeros(num_folder)
        paper_cote = serie + item
        page = NormalizeText.remove_leading_zeros(page)
        part_page = part_page[-1]
        return {"page_numeric_cote" : cote_page, "paper_cote": paper_cote, "archives" : archives, "serie" : serie, "item_in_serie" : item, "digitize_folder_num" : num_folder, "page": page, "part_in_double_page" : part_page}
    elif len(cote_parts) == 5:
        archives, serie, item, num_folder, page = cote_parts
        archives_ = archives.replace('FRAD','')
        archives = NormalizeText.remove_leading_zeros(archives_)
        item = NormalizeText.remove_leading_zeros(item)
        paper_cote = serie + item
        page = NormalizeText.remove_leading_zeros(page)
        num_folder = NormalizeText.remove_leading_zeros(num_folder)
        return {"page_numeric_cote" : cote_page, "paper_cote": paper_cote, "archives" : archives, "serie" : serie, "item_in_serie" : item, "digitize_folder_num" : num_folder, "page": page, "part_in_double_page" : ""}

pages_jsons = []
last_cover = ""
for _, row in df_pages.iterrows():
    if row["classe"] == "ets_couv":
        last_cover = row["id"]
    json_ = cote_page_parser(row["name"])
    json_["num_folder"] = row["dossier_cote"]
    json_["uuid"] = row["id"]
    json_["coordinates"] = row["coords"]
    json_["iiif_url"] = row["image_url"]
    json_["commune_folder"] = row["commune"]
    json_["ml_classe"] = row["classe"]
    json_["associated_page_cover_uuid"] = last_cover

    for ix, row2 in metadata_simple.iterrows():
        if row2["Cote"] == json_["num_folder"] and row2['Tag_groupe_doc'] != "AV_1822_B":
            json_["commune"] = row2["Commune"]
            json_["date_registre"] = row2["Date_doc"]
            json_["date_infos"] = row2["Date_infos"]
            json_["type_registre"] = row2["Tag_groupe_doc"]
            continue
    
    pages_jsons.append(json_)

from rdf_resources_utils import generate_source_rdf_resource

g = generate_source_rdf_resource(pages_jsons)

g.serialize(destination=f"{DATA}/rdf_gold/sources.ttl", format="turtle")

## 3. Create landmarks of type Plot, RecordPart for table lines and LandmarkRelations between Plots and Sections

print(len(JSONS))

### 3.1 Add page context infos to page JSON

for JSON in JSONS:
    page_uuid = JSON.replace(ROOT+'/',"").replace(".json","")
    print(page_uuid)
    with open(JSON) as f:
        page = json.load(f)

    page["context"] = {}
    page_json = next((item for item in pages_jsons if item.get("uuid") == page_uuid), None)
    page["context"]["section"] = "http://rdf.geohistoricaldata.org/id/landmark/" + page_json["associated_page_cover_uuid"]

    landmark_dict = next((item for item in landmarks_dict if item.get("uuid") == page_json["associated_page_cover_uuid"]), None)
    page["context"]["section_id"] = landmark_dict["identifier"]
    page["context"]["commune_uri"] = "http://rdf.geohistoricaldata.org/id/landmark/" + page_json["archives"] + '_' + page_json["commune_folder"]
    page["context"]["commune_folder"] = page_json["commune_folder"]
    page["context"]["commune"] = page_json["commune"]
    page["context"]["code_dept"] = page_json["archives"]
    page["context"]["iiif_url"] = page_json["iiif_url"][:page_json["iiif_url"].find('.jpg')+4] + '/info.json'
    page["context"]["page_numeric_cote"] = page_json["page_numeric_cote"]
    page["context"]["date"] = page_json["date_infos"]

    for line in page["entities"]:
        if "Ⓕ" in list(line.keys()) and 'interpreted_text' in list(line["Ⓕ"].keys()):
            if isinstance(line["Ⓕ"]['postcorrected_text'], int):
                line["Ⓕ"]['plot_id'] = landmark_dict["identifier"] + '-' + str(line["Ⓕ"]['postcorrected_text'])
            elif isinstance(line["Ⓕ"]['postcorrected_text'], float):
                line["Ⓕ"]['plot_id'] = landmark_dict["identifier"] + '-' + str(int(line["Ⓕ"]['postcorrected_text']))
            else:
                line["Ⓕ"]['plot_id'] = landmark_dict["identifier"] + '-' + line["Ⓕ"]['postcorrected_text']
        elif "Ⓕ" not in list(line.keys()):
            line["Ⓕ"] = {}
            line["Ⓕ"]['plot_id'] = landmark_dict["identifier"] + '-' + str("UNKNOWN")

    with open(JSON,'w', encoding='utf-8') as f:
        json.dump(page, f, ensure_ascii=False, indent=4)

### 3.2 Create the plot landmark and the table line

from rdf_resources_utils import generate_rdf_resource_plot_landmark, generate_rdf_resource_event

g = Graph()
events_ls = []
counter = 0
for JSON in JSONS:
    page_uuid = JSON.replace(ROOT+"/","").replace(".json","")
    with open(JSON) as f:
        page = json.load(f)
    generate_rdf_resource_plot_landmark(g, page, page_uuid)

    #Event
    uri_event = "http://rdf.geohistoricaldata.org/id/event/" + "CADASTRE_LHAY_" + str(page["context"]["date"])
    if uri_event not in events_ls:
        generate_rdf_resource_event(g, page)
        events_ls.append(uri_event)

page

g.serialize(destination=f"{DATA}/rdf_gold/initial-plots.ttl", format="turtle")

## 3.3 Create taxpayers

TAXPAYER_CSV = "/home/STual/DAN-cadastre/data/LHAY/corrected_updated/taxpayers_entities.csv"
taxpayers_df = pd.read_csv(TAXPAYER_CSV)
taxpayers_df

from linking_utils import LinkingUtils

# Function to generate the RDF resource with additional properties (address, activity, title)
def jsonprop_to_rdfprop(prop):
    CAD = Namespace("http://rdf.geohistoricaldata.org/def/cadastre#")
    if prop =="address" or prop == "addresses":
        return CAD.taxpayerAddress
    elif prop =="activity" or prop =="activities":
        return CAD.taxpayerActivity
    elif prop =="title" or prop =="titles":
        return CAD.taxpayerTitle
    elif prop =="familystatus":
        return CAD.taxpayerFamilyStatus

def normalize_whitespaces(label):
    label = re.sub('→',' ',label)
    label = re.sub('  ',' ',label)
    label = re.sub('[ ]+$','',label)
    return label

def generate_rdf_resource(df):
    """
    Create a simple RDF resource for each group with additional properties (address, activity, title, familystatus).
    """
    g = Graph()
    ADDR = Namespace("http://rdf.geohistoricaldata.org/def/address#")
    CAD = Namespace("http://rdf.geohistoricaldata.org/def/cadastre#")
    TAXPAYER = Namespace("http://rdf.geohistoricaldata.org/id/taxpayer/")
    g.bind("addr", ADDR)
    g.bind("cad", CAD)
    g.bind("taxpayer", TAXPAYER)

    rdf_entities = df.to_dict(orient='records')
    
    for rdf_entity in rdf_entities:
        
        uri = URIRef("http://rdf.geohistoricaldata.org/id/taxpayer/" + rdf_entity['taxpayer_uuid'])
        g.add((uri, RDF.type, CAD.Taxpayer))
    
        # Name and firstnames
        if pd.notna(rdf_entity['name']) and pd.notna(rdf_entity['firstnames']) and pd.notna(rdf_entity['familystatus']):
            label = rdf_entity['name'] + ' ' + rdf_entity['firstnames'] + ' ' + rdf_entity['familystatus']
            g.add((uri, CAD.taxpayerLabel, Literal(rdf_entity['name'])))
            g.add((uri, CAD.taxpayerFirstnames, Literal(rdf_entity['firstnames'])))
            g.add((uri, CAD.taxpayerFamilyStatus, Literal(rdf_entity['familystatus'])))
        elif pd.notna(rdf_entity['name']) and pd.notna(rdf_entity['firstnames']) and pd.isna(rdf_entity['familystatus']):
            label = rdf_entity['name'] + ' ' + rdf_entity['firstnames']
            g.add((uri, CAD.taxpayerLabel, Literal(rdf_entity['name'])))
            g.add((uri, CAD.taxpayerFirstnames, Literal(rdf_entity['firstnames'])))
        elif pd.notna(rdf_entity['name']) and pd.isna(rdf_entity['firstnames']) and pd.notna(rdf_entity['familystatus']):
            label = rdf_entity['name'] + ' ' + rdf_entity['familystatus']
            g.add((uri, CAD.taxpayerLabel, Literal(rdf_entity['name'])))
            g.add((uri, CAD.taxpayerFamilyStatus, Literal(rdf_entity['familystatus'])))
        else:
            label = rdf_entity['name']
            g.add((uri, CAD.taxpayerLabel, Literal(rdf_entity['name'])))

        g.add((uri, RDFS.label, Literal(normalize_whitespaces(label))))
    
        # Adding address, activity, and title properties
        for property_name in ['addresses', 'activities', 'titles']:
            if pd.notna(rdf_entity[property_name]):
                g.add((uri, jsonprop_to_rdfprop(property_name), Literal(rdf_entity[property_name][0].title())))

        if pd.notna(rdf_entity['taxpayer_num']):
            g.add((uri, CAD.hasIndexNumber, Literal(str(int(rdf_entity['taxpayer_num'])))))

    return g

# Generate RDF resource
graph = generate_rdf_resource(taxpayers_df)

# Print the RDF graph in Turtle format
graph.serialize(destination=f"{DATA}/rdf_gold/taxpayers.ttl", format="turtle")

## 3.4 Create other landmarks

LANDMARKS_CSV = "/home/STual/DAN-cadastre/data/LHAY/corrected_updated/landmark_entities.csv"
lieudit_df = pd.read_csv(LANDMARKS_CSV)
lieudit_df

def generate_rdf_resource(df):
    """
    Create a simple RDF resource for each group with additional properties (address, activity, title).
    """
    g = Graph()
    ADDR = Namespace("http://rdf.geohistoricaldata.org/def/address#")
    CAD = Namespace("http://rdf.geohistoricaldata.org/def/cadastre#")
    LR = Namespace("http://rdf.geohistoricaldata.org/id/landmarkRelation/")
    LTYPE = Namespace("http://rdf.geohistoricaldata.org/id/codes/address/landmarkType/")
    LANDMARK = Namespace("http://rdf.geohistoricaldata.org/id/landmark/")
    CAD_LTYPE = Namespace("http://rdf.geohistoricaldata.org/id/codes/cadastre/landmarkType/")
    LRTYPE = Namespace("http://rdf.geohistoricaldata.org/id/codes/address/landmarkRelationType/")
    g.bind("addr", ADDR)
    g.bind("cad", CAD)
    g.bind("landmark", LANDMARK)
    g.bind("cad_ltype", CAD_LTYPE)
    g.bind("ltype", LTYPE)
    g.bind("lrtype", LRTYPE)

    rdf_entities = df.to_dict(orient='records')
    
    for rdf_entity in rdf_entities:
        uri = URIRef(LANDMARK + rdf_entity['landmark_uuid'])
        g.add((uri, RDF.type, ADDR.Landmark))
        g.add((uri, ADDR.isLandmarkType, LTYPE.District))

        #uri_lr = URIRef(BNode().n3())
        p1 = uri.replace("http://rdf.geohistoricaldata.org/id/landmark/","")
        uri_lr = URIRef(LR + p1 + '_' + DEP + '_' + COMMUNE)
        g.add((uri_lr, RDF.type, ADDR.LandmarkRelation))
        g.add((uri_lr, ADDR.isLandmarkRelationType, LRTYPE.Within))
        g.add((uri_lr, ADDR.locatum, uri))
        g.add((uri_lr, ADDR.relatum, URIRef(LANDMARK + DEP + '_' + COMMUNE)))

        label = rdf_entity['rdfs:label']
        label = re.sub('→',' ',label)
        label = re.sub('  ',' ',label)
        label = re.sub('[ ]+$','',label)
        g.add((uri, RDFS.label, Literal(label)))

    return g

# Generate RDF resource
graph = generate_rdf_resource(lieudit_df)

# Print the RDF graph in Turtle format
graph.serialize(destination=f"{DATA}/rdf_gold/lieu-dit.ttl", format="turtle")
